BUG: Fix comment ternary operator
Just because my static analyzer keeps bugging me about it really.

If you want to add a test for this: The broken path would be passing
in an empty comment, because it would read a byte too many.

There are more issues here: These are all `char` and not unicode
as they should be...  Apparently, there are probably no tests that
use unicode separators, even though the code should handle them just
fine.