WIP: Rework numpy dtypes into everything
This also reorganizes a bit, and I just deleted the currently
unused `from_filename`, that seemed easier, since otherwise it
just means updating duplicated things.

I modified the type structure to rely on itemsize rather than codes.
That currently breaks some errors (because it has the incorrect
types for the fields).

In the current stage it may still crash down in flames, but it is
a start, and probably fairly far along.